Refactoring: moved regressor architectures into deeplearning. Moved base regressor to base/estimators/_regressor.py. Tests pass.
