+++ sktime_dl/deeplearning/cnn/__init__.py
@@ -1,1 +1,2 @@
-from ._cnn import CNNClassifier
+from ._classifier import CNNClassifier
+from ._regressor import CNNRegressor
