+++ sktime_dl/deeplearning/mlp/_classifier.py
@@ -1,18 +1,13 @@
-__author__ = "James Large, Withington"
+__author__ = "James Large"
 
 import keras
 import numpy as np
-import pandas as pd
 
-from sklearn.base import RegressorMixin
-
-from sktime.utils.validation.supervised import validate_X, validate_X_y
-
-from sktime_dl.regressors.deeplearning._base import BaseDeepRegressor
+from sktime_dl.base.estimators import BaseDeepClassifier
 from sktime_dl.networks.deeplearning import MLPNetwork
 
 
-class MLPRegressor(BaseDeepRegressor, RegressorMixin):
+class MLPClassifier(BaseDeepClassifier):
     """Multi Layer Perceptron (MLP).
 
     Adapted from the implementation from Fawaz et. al
@@ -32,16 +27,16 @@ class MLPRegressor(BaseDeepRegressor, RegressorMixin):
     """
 
     def __init__(self,
-                 nb_epochs=2000,
+                 nb_epochs=5000,
                  batch_size=16,
 
                  random_seed=0,
                  verbose=False,
-                 model_name="mlp_regressor",
+                 model_name="mlp",
                  model_save_directory=None):
         '''
         :param nb_epochs: int, the number of epochs to train the model
-        :param batch_size: int, the number of samples per gradient update.
+        :param batch_size: int, specifying the length of the 1D convolution window
         :param random_seed: int, seed to any needed random actions
         :param verbose: boolean, whether to output extra information
         :param model_name: string, the name of this model for printing and file writing purposes
@@ -68,22 +63,25 @@ class MLPRegressor(BaseDeepRegressor, RegressorMixin):
         self.random_seed = random_seed
         self.random_state = np.random.RandomState(self.random_seed)
 
-    def build_model(self, input_shape, **kwargs):
+    def build_model(self, input_shape, nb_classes, **kwargs):
         """
         Construct a compiled, un-trained, keras model that is ready for training
         ----------
         input_shape : tuple
             The shape of the data fed into the input layer
+        nb_classes: int
+            The number of classes, which shall become the size of the output layer
         Returns
         -------
         output : a compiled Keras Model
         """
         network = MLPNetwork(self.random_seed)
         input_layer, output_layer = network.build_network(input_shape, **kwargs)
-        output_layer = keras.layers.Dense(units=1)(output_layer)
+        output_layer = keras.layers.Dense(nb_classes, activation='softmax')(output_layer)
 
         model = keras.models.Model(inputs=input_layer, outputs=output_layer)
-        model.compile(loss='mean_squared_error', optimizer=keras.optimizers.Adadelta(),
+
+        model.compile(loss='categorical_crossentropy', optimizer=keras.optimizers.Adadelta(),
                       metrics=['accuracy'])
 
         reduce_lr = keras.callbacks.ReduceLROnPlateau(monitor='loss', factor=0.5, patience=200, min_lr=0.1)
@@ -94,12 +92,12 @@ class MLPRegressor(BaseDeepRegressor, RegressorMixin):
 
     def fit(self, X, y, input_checks=True, **kwargs):
         """
-        Build the regressor on the training set (X, y)
+        Build the classifier on the training set (X, y)
         ----------
         X : array-like or sparse matrix of shape = [n_instances, n_columns]
             The training input samples.  If a Pandas data frame is passed, column 0 is extracted.
         y : array-like, shape = [n_instances]
-            The regression values.
+            The class labels.
         input_checks: boolean
             whether to check the X and y parameters
         Returns
@@ -107,16 +105,18 @@ class MLPRegressor(BaseDeepRegressor, RegressorMixin):
         self : object
         """
         X = self.check_and_clean_data(X, y, input_checks=input_checks)
+
+        y_onehot = self.convert_y(y)
         self.input_shape = X.shape[1:]
 
-        self.batch_size = int(max(1, min(X.shape[0] / 10, self.batch_size)))
+        self.batch_size = int(min(X.shape[0] / 10, self.batch_size))
 
-        self.model = self.build_model(self.input_shape)
+        self.model = self.build_model(self.input_shape, self.nb_classes)
 
         if self.verbose:
             self.model.summary()
 
-        self.history = self.model.fit(X, y, batch_size=self.batch_size, epochs=self.nb_epochs,
+        self.history = self.model.fit(X, y_onehot, batch_size=self.batch_size, epochs=self.nb_epochs,
                                       verbose=self.verbose, callbacks=self.callbacks)
 
         self.save_trained_model()
