--- sktime_dl/regressors/deeplearning/__init__.py
@@ -1,2 +1,0 @@
-from ._cnn import CNNRegressor
-from ._mlp import MLPRegressor
